Update tests for Aphrodite 1.2.0
In Aphrodite 1.2.0, we replaced the hashing algorithm for better
performance. This had the side-effect of changing the class names that
Aphrodite generates. For now I'm just going to update these to match.

This is an implementation detail of Aphrodite, so there's really nothing
else that we need to do to maintain compatibility.

Fixes #7.