Improve tree-shakeability of Aphrodite
I originally added support for Aphrodite v2 in a backwards-compatible
way, but had to resort to using CommonJS in a try/catch to make it work.
This will make Aphrodite less tree-shakeable than using the named
export, so I want to get things on that now. This means we have to drop
support for v1, making this a breaking change.

v2 now uses ES modules with named exports, and exposes flushToStyleTag
explicitly instead of requiring a deep import.

More info: https://twitter.com/lencioni/status/974161112110280704