Replace array-flatten with array.prototype.flat
This is the shim package for the standard Array.prototype.flat, so this
will use the native function in environments where it is available.

Since we might have deeply nested arrays, I decided to pass flat a depth
of Infinity. Looking at the test suite, this seems to be the correct way
to accomplish this:

  https://github.com/es-shims/Array.prototype.flat/blob/0bb5b417/test/tests.js#L19